requirements.txt
langchain
langchain_core
langchain_community
langchain_classic
langchain_chroma
streamlit
langchain_huggingface
langchain_groq
pypdf
python-dotenv
sentence_transformers